- client: fix a bug that caused GPU apps to fail on startup     in the following situation.     Suppose GPU job A is running,     and GPU job B (which has never run before) preempts it.     Then     1) an ACTIVE_TASK and slot dir are allocated for B (slot dir is empty)     2) A is told to quit; B isn't started until A exits         (because they're GPU jobs)     3) before A exits, B is suspended     4) eventually B is started.         Its scheduler state is SUSPENDED rather than UNINITIALIZED,         so the scheduler thinks it slot dir has already been populated.         But it hasn't, and when the app starts it fails immediately         because no files are there.
    Fix: decide whether a slot dir needs to be populated by
    seeing whether it's empty,
    rather than looking at the scheduler state of the task.


svn path=/trunk/boinc/; revision=22206

+++ client/http_curl.cpp
@@ -1030,8 +1030,8 @@ void HTTP_OP::handle_messages(CURLMsg *pcurlMsg) {
             if (nread != dSize) {
                 if (log_flags.http_debug) {
                     msg_printf(NULL, MSG_INFO,
-                        "[http] post output file read failed %ld",
-                        nread
+                        "[http] post output file read failed %d",
+                        (int)nread
                     );
                 }
             }
